ci(windows): Windows tient la même ligne que Linux (/WX)
Le job Windows compilait en `/W4` avec les avertissements VISIBLES mais non
fatals, parce que le jeu d'avertissements de MSVC n'avait jamais été mesuré sur
cet arbre. Il l'est maintenant : 134 sites distincts.

102 d'entre eux sont dans `stb_vorbis.c`. Il est vendu, et déjà silencé côté GCC
et clang par un bloc de pragmas — jamais côté MSVC. Le `.c` étant inclus dans
`AudioDevice.cpp`, ses avertissements étaient imputés à CETTE unité de
traduction, exactement la raison pour laquelle les deux autres compilateurs
avaient besoin du bloc. Un `warning(push, 0)` autour de l'inclusion et les 102
disparaissent. Même traitement pour `miniaudio.h`, y compris là où
`CassetteDevice.h` le tire dans tout ce qui l'inclut.

Restent 30 sites dans le code POM1, tous CORRIGÉS et non désactivés :

  C4244  conversions rétrécissantes rendues explicites par static_cast —
         M6502, TMS9918, HgrConvert, SidTrackerEditor, six tests. Dans
         CassetteDevice, `std::llround` devient `std::round` : la valeur était
         de toute façon ramenée dans le domaine double juste après, et llround
         était à la fois un rétrécissement et un comportement indéfini sur les
         valeurs hors-bornes que cette arithmétique existe pour survivre.
  C4456  quatre déclarations masquantes renommées (CliDispatcher,
  C4457  MainWindow_Presets, CodeBench, hgr_convert_smoke_test).
  C4996  inet_ntoa → inet_ntop dans TerminalCard et TelemetryPort. Déprécié sur
         Windows, et il rend de toute façon un pointeur vers un tampon statique
         partagé sur toutes les plateformes.

UN DE CES MASQUAGES ÉTAIT UN VRAI DÉFAUT. Dans `hgr_convert_smoke_test`, le bloc
« les bandes de letterbox restent noires » déclarait un `img` local de 100×192
puis passait au convertisseur… l'`img` extérieur de 280×192. Il testait donc le
letterbox sur une source qui n'en a pas. Corrigé, le test passe toujours :
l'assertion tient sur l'entrée qu'elle prétendait utiliser. C'est exactement ce
qu'achète une porte d'avertissements, et ce qu'aucune relecture n'avait vu en
un an.

Le nom de remplacement n'est pas `small` non plus — <windows.h> (rpcndr.h) en
fait une macro pour `char`, ce qui aurait déplacé le problème sur la seule
plateforme qu'on essaie de rendre verte.

Vérifié d'ici : suite complète verte (120/120), et un build `-DPOM1_WERROR=ON`
complet sous clang sans une seule erreur. Le verdict MSVC appartient au job
Windows, que ce commit passe en `-DPOM1_WERROR=ON`.

+++ src/M6502.cpp
@@ -547,7 +547,7 @@ void M6502::ROR_A(void)
     else
         statusRegister &= ~M6502::Status::C;
 
-    accumulator = tmp >> 1;
+    accumulator = static_cast<uint8_t>(tmp >> 1);
     setStatusRegisterNZ(accumulator);
 }
 
@@ -1282,7 +1282,7 @@ void M6502::hardReset(void)
 
     if (memory != nullptr) {
         for (int i = 0x100; i <= 0x1FF; i++) {
-            memory->memWrite(i, 0x00);
+            memory->memWrite(static_cast<uint16_t>(i), 0x00);
         }
     }
